+++ app/src/main/java/org/simple/clinic/main/TypedPreference.kt
@@ -23,5 +23,7 @@ annotation class TypedPreference(val value: Type) {
     LastQuestionnairePullToken,
     LastQuestionnaireResponsePullToken,
     DataProtectionConsent,
+    LastPatientAttributePullToken,
+    LastCVDRiskPullToken,
   }
 }