+++ app/src/test/java/org/simple/clinic/medicalhistory/newentry/NewMedicalHistoryUiRendererTest.kt
@@ -1,9 +1,9 @@
 package org.simple.clinic.medicalhistory.newentry
 
+import org.junit.Test
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoMoreInteractions
-import org.junit.Test
 import org.simple.clinic.appconfig.Country
 import org.simple.clinic.facility.FacilityConfig
 import org.simple.clinic.medicalhistory.Answer.No
@@ -14,6 +14,7 @@ import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.DiagnosedWithHype
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HasHadAHeartAttack
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HasHadAKidneyDisease
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HasHadAStroke
+import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IsSmoker
 import org.simple.sharedTestCode.TestData
 import java.util.UUID
 
@@ -42,7 +43,7 @@ class NewMedicalHistoryUiRendererTest {
       )
 
   private val country = TestData.country(isoCountryCode = Country.INDIA)
-  private val defaultModel = NewMedicalHistoryModel.default(country)
+  private val defaultModel = NewMedicalHistoryModel.default(country, false)
 
   private val ui = mock<NewMedicalHistoryUi>()
   private val uiRenderer = NewMedicalHistoryUiRenderer(ui)
@@ -69,6 +70,7 @@ class NewMedicalHistoryUiRendererTest {
     verify(ui).hideDiabetesDiagnosisView()
     verify(ui).showDiabetesHistorySection()
     verify(ui).renderAnswerForQuestion(DiagnosedWithDiabetes, Unanswered)
+    verify(ui).hideCurrentSmokerQuestion()
     verifyNoMoreInteractions(ui)
   }
 
@@ -162,7 +164,7 @@ class NewMedicalHistoryUiRendererTest {
   fun `when patient has hypertension and country is not from india, then don't show hypertension treatment question`() {
     // given
     val bangladesh = TestData.country(isoCountryCode = Country.BANGLADESH)
-    val model = NewMedicalHistoryModel.default(country = bangladesh)
+    val model = NewMedicalHistoryModel.default(country = bangladesh, false)
         .currentFacilityLoaded(facilityWithDiabetesManagementEnabled)
         .answerChanged(DiagnosedWithHypertension, Yes)
 
@@ -228,7 +230,7 @@ class NewMedicalHistoryUiRendererTest {
   fun `when diabetes management is enabled and patient has diabetes and is not from india, then don't show diabetes treatment question`() {
     // given
     val bangladesh = TestData.country(isoCountryCode = Country.BANGLADESH)
-    val model = NewMedicalHistoryModel.default(country = bangladesh)
+    val model = NewMedicalHistoryModel.default(country = bangladesh, false)
         .currentFacilityLoaded(facilityWithDiabetesManagementEnabled)
         .answerChanged(DiagnosedWithDiabetes, Yes)
 
@@ -269,9 +271,65 @@ class NewMedicalHistoryUiRendererTest {
     verifyNoMoreInteractions(ui)
   }
 
+  @Test
+  fun `when show smoker question is enabled, then show current smoker question`() {
+    // given
+    val model = NewMedicalHistoryModel.default(country, true)
+        .answerChanged(DiagnosedWithHypertension, Unanswered)
+        .answerChanged(HasHadAHeartAttack, Yes)
+        .answerChanged(HasHadAStroke, No)
+        .answerChanged(HasHadAKidneyDisease, Unanswered)
+        .answerChanged(IsSmoker, No)
+
+    // when
+    uiRenderer.render(model)
+
+    // then
+    verify(ui).renderDiagnosisAnswer(DiagnosedWithHypertension, Unanswered)
+    verify(ui).hideHypertensionTreatmentQuestion()
+    verify(ui).renderAnswerForQuestion(HasHadAHeartAttack, Yes)
+    verify(ui).renderAnswerForQuestion(HasHadAStroke, No)
+    verify(ui).renderAnswerForQuestion(HasHadAKidneyDisease, Unanswered)
+    verify(ui).hideNextButtonProgress()
+    verify(ui).hideDiabetesDiagnosisView()
+    verify(ui).showDiabetesHistorySection()
+    verify(ui).renderAnswerForQuestion(DiagnosedWithDiabetes, Unanswered)
+    verify(ui).showCurrentSmokerQuestion()
+    verify(ui).renderAnswerForQuestion(IsSmoker, No)
+    verifyNoMoreInteractions(ui)
+  }
+
+  @Test
+  fun `when show smoker question is disabled, then hide current smoker question`() {
+    // given
+    val model = NewMedicalHistoryModel.default(country, false)
+        .answerChanged(DiagnosedWithHypertension, Unanswered)
+        .answerChanged(HasHadAHeartAttack, Yes)
+        .answerChanged(HasHadAStroke, No)
+        .answerChanged(HasHadAKidneyDisease, Unanswered)
+
+    // when
+    uiRenderer.render(model)
+
+    // then
+    verify(ui).renderDiagnosisAnswer(DiagnosedWithHypertension, Unanswered)
+    verify(ui).hideHypertensionTreatmentQuestion()
+    verify(ui).renderAnswerForQuestion(HasHadAHeartAttack, Yes)
+    verify(ui).renderAnswerForQuestion(HasHadAStroke, No)
+    verify(ui).renderAnswerForQuestion(HasHadAKidneyDisease, Unanswered)
+    verify(ui).hideNextButtonProgress()
+    verify(ui).hideDiabetesDiagnosisView()
+    verify(ui).showDiabetesHistorySection()
+    verify(ui).renderAnswerForQuestion(DiagnosedWithDiabetes, Unanswered)
+    verify(ui).hideCurrentSmokerQuestion()
+    verifyNoMoreInteractions(ui)
+  }
+
+
   private fun verifyImplicitRenders() {
     verify(ui).renderAnswerForQuestion(HasHadAHeartAttack, Unanswered)
     verify(ui).renderAnswerForQuestion(HasHadAStroke, Unanswered)
     verify(ui).renderAnswerForQuestion(HasHadAKidneyDisease, Unanswered)
+    verify(ui).hideCurrentSmokerQuestion()
   }
 }